+++ components/Layout/index.tsx
@@ -1,2 +1,2 @@
 export { default as HomeLayout } from "./HomeLayout";
-export { default as Layout } from "./Layout";
+export { default as Layout } from "./PageLayout";
