+++ pages/[city]/index.tsx
@@ -1,16 +1,51 @@
 import cities, { ICategoryData } from "@mauriciorobayo/pyptron";
+import TheMoneytizer from "components/ads/the-moneytizer";
 import Vidverto from "components/ads/vidverto";
+import { Aside } from "components/Aside";
+import CTA from "components/call-to-action";
 import CategoriesList from "components/categories-list";
 import { getLocalLongDateString } from "components/date/utils";
-import { Layout } from "components/Layout";
-import { cityOptions, CityOptions } from "components/select/utils";
+import { Header } from "components/Header";
 import { AMERICA_BOGOTA, CityType, dateParts, isCity } from "lib/utils";
 import { GetStaticPaths, GetStaticProps } from "next";
 import Link from "next/link";
+import React from "react";
 import styled from "styled-components";
 
+const Main = styled.main`
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  padding: 0 1rem;
+`;
+
 const StyledVidverto = styled(Vidverto)`
-  margin: 1rem 0 0;
+  margin-top: 1rem;
+`;
+
+const Page = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  width: 100%;
+`;
+
+const MegaBanner = styled(TheMoneytizer).attrs({
+  formatType: "MEGABANNER",
+})`
+  margin: 2rem auto 0;
+`;
+
+const MegaBannerBottom = styled(TheMoneytizer).attrs({
+  formatType: "MEGABANNER_BOTTOM",
+})`
+  margin: 0 auto 2rem;
+`;
+
+const RecommendedContent = styled(TheMoneytizer).attrs({
+  formatType: "RECOMMENDED_CONTENT",
+})`
+  margin: 2rem auto;
 `;
 
 const StyledCategoriesList = styled(CategoriesList)`
@@ -21,7 +56,6 @@ type CityProps = {
   categories: ICategoryData[];
   cityName: string;
   citySlug: CityType;
-  selectOptions: CityOptions;
   currentDate: number;
 };
 
@@ -30,51 +64,52 @@ export default function City({
   cityName,
   citySlug,
   currentDate,
-  selectOptions,
 }: CityProps) {
-  const title = `Pico y placa ${cityName}`;
   const date = new Date(currentDate);
-
-  const aside = (
-    <section>
-      <h4>
-        Pico y placa vigente en
-        {cityName}
-      </h4>
-      <p>
-        Las siguientes son las medidas de restricción vehicular vigentes para{" "}
-        {cityName} durante el mes de{" "}
-        {getLocalLongDateString().split(" ").slice(3).join(" ")}, de acuerdo con
-        lo establecido por la Alcaldía de
-        {cityName}:
-      </p>
-      <ul>
-        {categories.map(({ name: categoryName, slug: categorySlug }) => (
-          <li key={categoryName}>
-            <Link href={`/${citySlug}/${categorySlug}`}>
-              <a>{categoryName}</a>
-            </Link>
-          </li>
-        ))}
-      </ul>
-    </section>
-  );
+  const title = `Pico y placa ${cityName}`;
 
   return (
-    <Layout
-      aside={aside}
-      cityName={cityName}
-      date={date}
-      selectOptions={selectOptions}
-      title={title}
-    >
-      <StyledVidverto />
-      <StyledCategoriesList
-        categories={categories}
-        citySlug={citySlug}
-        date={date}
-      />
-    </Layout>
+    <>
+      <Page>
+        <MegaBanner />
+        <Header date={date} title={title} />
+        <Main>
+          <StyledVidverto />
+          <StyledCategoriesList
+            categories={categories}
+            citySlug={citySlug}
+            date={date}
+          />
+        </Main>
+        <MegaBannerBottom />
+        <CTA />
+      </Page>
+      <Aside>
+        <section>
+          <h4>
+            Pico y placa vigente en
+            {cityName}
+          </h4>
+          <p>
+            Las siguientes son las medidas de restricción vehicular vigentes
+            para {cityName} durante el mes de{" "}
+            {getLocalLongDateString().split(" ").slice(3).join(" ")}, de acuerdo
+            con lo establecido por la Alcaldía de
+            {cityName}:
+          </p>
+          <ul>
+            {categories.map(({ name: categoryName, slug: categorySlug }) => (
+              <li key={categoryName}>
+                <Link href={`/${citySlug}/${categorySlug}`}>
+                  <a>{categoryName}</a>
+                </Link>
+              </li>
+            ))}
+          </ul>
+        </section>
+        <RecommendedContent />
+      </Aside>
+    </>
   );
 }
 
@@ -107,7 +142,6 @@ export const getStaticProps: GetStaticProps = async ({ params }) => {
       cityName,
       citySlug,
       currentDate: date.getTime(),
-      selectOptions: cityOptions(),
     },
   };
 };
